Flatten nested combined link functions.
When combining a list of link functions, detect if any of them are the
result of combining another list of link functions and, if so, flatten
them out into a single function.

This will reduce the layers of indirection caused when combining together
link functions produced for the same level of the tree.