Allow directives and filters to be registered via methods.
Registering them in the initial call is bothersome because they will
usually depend on the parse, interpolate or compile methods of the
compiler object and thus there is a chicken-and-egg problem.

Instead, we now register them separately, and methods are provided to
register sets of directives/filters in one shot under the expectation that
a large application will have several sets that come from different
libraries that all need to get registered on the compiler.