Support wrapper components.
A wrapper component is a component directive that transcludes its original
content into some point inside its template. This is implemented by
keeping track of the active content link function as we pass from one
link function to another, so this is a new standard parameter on all
node-appending link functions.