Flesh out the decorator directive API.
Decorator directives have pre- and post-link functions that make them
ideal hosts for services, since they can establish link-time state during
pre-link (which runs before child nodes are linked) and then clean up
that state during post-link (which runs after all of the child nodes are
complete).

There is a remaining wart here in that this mechanism doesn't work
properly with component directives that produce more than one node. This
will be fixed with a change to the component directive API in a separate
commit.