Add thi-switch directive construct.
This is the first multi-directive construct created using the services
mechanism. thi-switch establishes a context where the value of a
particular expression can be switched on, and then child directives
thi-switch-when and thi-switch-else use that context to show only elements
that relate to the active state.

In this initial version the 'when' expressions are required to be
constant, but this may be relaxed in a later revision.